Projects/saas-starter/backend/requirements.txt
Django==4.2.10
stripe==10.0.0
PyJWT==2.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cors-headers==4.3.1
requests==2.31.0
dj-database-url==2.1.0
boto3==1.28.72
django-storages==1.13.1
sentry-sdk==1.27.0
django-ratelimit==4.1.0
redis==5.0.1
django-redis==5.4.0
djangorestframework==3.14.0
openai==0.28.1